livinit_pipeline-main/requirements.txt
fastapi
uvicorn
python-multipart
mangum
python-dotenv>=1.0.1
supabase>=2.21.1
matplotlib>=3.8.4
usd-core>=24.11
google-genai>=1.0.0
langgraph>=0.2.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
numpy>=1.21.0
openai>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0
shapely>=2.0.0
imageio>=2.9.0
pydantic>=2.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
opencv-python-headless>=4.5.0
pillow>=9.0.0
requests>=2.28.0
trimesh>=3.9.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch
timm
einops
